feat: supporting variety of rootfs types
- lower dir can be provided by user wihout requiring squashfs
- system can start without providint folder, single binnary can be attached and started

+++ pkg/container/rootfs.go
@@ -3,6 +3,7 @@ package container
 import (
 	"fmt"
 	"os"
+	"strings"
 
 	"github.com/ahmetozer/sandal/pkg/config"
 	"golang.org/x/sys/unix"
@@ -10,9 +11,14 @@ import (
 
 func MountRootfs(c *config.Config) error {
 	// Mount overlay filesystem
-	squasfsMount, err := mountSquashfsFile(c)
-	if err != nil {
-		return fmt.Errorf("mounting squashfs file: %s", err)
+
+	if c.SquashfsFile != "" {
+		squasfsMount, err := mountSquashfsFile(c)
+		if err != nil {
+			return fmt.Errorf("mounting squashfs file: %s", err)
+		}
+		// this will last item of c.LowerDirs and lowest priority
+		c.LowerDirs = append(c.LowerDirs, squasfsMount)
 	}
 
 	changeDir, err := createChangeDir(c)
@@ -24,10 +30,25 @@ func MountRootfs(c *config.Config) error {
 		return fmt.Errorf("creating workdir: %s", err)
 	}
 
-	options := fmt.Sprintf("lowerdir=%s,upperdir=%s,workdir=%s", squasfsMount, changeDir.uppper, changeDir.work)
-	err = unix.Mount("overlay", c.RootfsDir, "overlay", 0, options)
-	if err != nil {
-		return fmt.Errorf("overlay: %s", err)
+	if len(c.LowerDirs) == 0 {
+		if len(c.Volumes) == 0 {
+			return fmt.Errorf("no lower dir is provided")
+		}
+	} else {
+		// check folder is exist
+		for _, folder := range c.LowerDirs {
+			if _, err := os.Stat(folder); err != nil {
+				return fmt.Errorf("folder %s is not exist: %e", folder, err)
+			}
+		}
+	}
+
+	if len(c.LowerDirs) != 0 {
+		options := fmt.Sprintf("lowerdir=%s,upperdir=%s,workdir=%s", strings.Join(c.LowerDirs, ":"), changeDir.uppper, changeDir.work)
+		err = unix.Mount("overlay", c.RootfsDir, "overlay", 0, options)
+		if err != nil {
+			return fmt.Errorf("overlay: %s", err)
+		}
 	}
 	return nil
 
@@ -35,24 +56,28 @@ func MountRootfs(c *config.Config) error {
 
 func UmountRootfs(c *config.Config) []error {
 	errors := []error{}
-	err := umountSquashfsFile(c)
-	if err != nil {
-		errors = append(errors, err)
-	}
-
-	err = unix.Unmount(c.RootfsDir, 0)
-	if err != nil {
-		if !os.IsNotExist(err) {
-			errors = append(errors, err)
-		}
-	}
-	err = os.Remove(c.RootfsDir)
-	if err != nil {
-		if !os.IsNotExist(err) {
+	var err error
+	if c.SquashfsFile != "" {
+		err = umountSquashfsFile(c)
+		if err != nil {
 			errors = append(errors, err)
 		}
 	}
 
+	if len(c.LowerDirs) != 0 || c.SquashfsFile != "" {
+		err = unix.Unmount(c.RootfsDir, 0)
+		if err != nil {
+			if !os.IsNotExist(err) {
+				errors = append(errors, err)
+			}
+		}
+		err = os.Remove(c.RootfsDir)
+		if err != nil {
+			if !os.IsNotExist(err) {
+				errors = append(errors, err)
+			}
+		}
+	}
 	if c.ChangeDir == "" && c.TmpSize != 0 {
 		err = unix.Unmount(defaultChangeRoot(c), 0)
 		if err != nil {
@@ -62,10 +87,13 @@ func UmountRootfs(c *config.Config) []error {
 		}
 	}
 
-	err = DetachLoopDevice(c.LoopDevNo)
-	if err != nil {
-		errors = append(errors, err)
+	if c.SquashfsFile != "" {
+		err = DetachLoopDevice(c.LoopDevNo)
+		if err != nil {
+			errors = append(errors, err)
+		}
 	}
+
 	if len(errors) == 0 {
 		return nil
 	}
